requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
mistralai==0.1.2
python-dotenv==1.0.1
Pillow==10.0.0
streamlit>=1.33.0 